requirements.txt
fastapi
uvicorn
sqlmodel
psycopg[binary]
fastapi-jwt-auth3
passlib
passlib[bcrypt]
python-dotenv
google-auth
pypdf
pymupdf
pypdf2
python-multipart
stripe
asyncpg
orjson
gunicorn
email_validator
psycopg_pool
